+++ ide-common/src/main/kotlin/org/digma/intellij/plugin/posthog/ActivityMonitor.kt
@@ -34,7 +34,7 @@ class ActivityMonitor(project: Project) : Disposable {
             return project.getService(ActivityMonitor::class.java)
         }
     }
-
+    private val installStatusPropertyName: String = "install status"
     private val userId: String = UserId.userId
     private val isDevUser: Boolean = UserId.isDevUser
     private val latestUnknownRunConfigTasks = mutableMapOf<String, Instant>()
@@ -73,9 +73,9 @@ class ActivityMonitor(project: Project) : Disposable {
         val mutableDetails: MutableMap<String, Any> = mutableMapOf()
         mutableDetails.putAll(details)
 
-        mutableDetails["firstTimeInsightReceived"] = PersistenceService.getInstance().state.firstTimeInsightReceived
-        mutableDetails["firstTimeAssetsReceived"] = PersistenceService.getInstance().state.firstTimeAssetsReceived
-        mutableDetails["firstTimeRecentActivityReceived"] = PersistenceService.getInstance().state.firstTimeRecentActivityReceived
+        mutableDetails["firstTimeInsightReceived"] = PersistenceService.getInstance().isFirstTimeInsightReceived()
+        mutableDetails["firstTimeAssetsReceived"] = PersistenceService.getInstance().isFirstTimeAssetsReceived()
+        mutableDetails["firstTimeRecentActivityReceived"] = PersistenceService.getInstance().isFirstTimeRecentActivityReceived()
 
         postHog?.capture(
             userId,
@@ -242,6 +242,14 @@ class ActivityMonitor(project: Project) : Disposable {
                 )
             )
         }
+    }
+
+
+    fun registerFatalError(details: MutableMap<String, String>) {
+        capture(
+            "fatal error",
+            details
+        )
     }
 
 
@@ -475,7 +483,6 @@ class ActivityMonitor(project: Project) : Disposable {
         registerUserAction("Clicked on view dashboard")
     }
 
-
     //todo: remove at some point
     fun registerFirstTimePluginLoaded() {
         postHog?.capture(userId, "plugin first-loaded")
@@ -491,11 +498,27 @@ class ActivityMonitor(project: Project) : Disposable {
 
     fun registerPluginUninstalled(): String {
         postHog?.capture(userId, "plugin uninstalled")
+        if(PersistenceService.getInstance().hasEmail()){
+            postHog?.capture(userId, "registered user uninstalled")
+        }
+        postHog?.set(
+            userId, mapOf(
+                installStatusPropertyName to "Uninstalled"
+            )
+        )
         return userId
     }
 
     fun registerPluginDisabled(): String {
         postHog?.capture(userId, "plugin disabled")
+        if(PersistenceService.getInstance().hasEmail()){
+            postHog?.capture(userId, "registered user disabled")
+        }
+        postHog?.set(
+            userId, mapOf(
+                installStatusPropertyName to "Disabled"
+            )
+        )
         return userId
     }
 
@@ -608,7 +631,8 @@ class ActivityMonitor(project: Project) : Disposable {
                 "ide.version" to ideVersion,
                 "ide.build" to ideBuildNumber,
                 "plugin.version" to pluginVersion,
-                "user.type" to if (isDevUser) "internal" else "external"
+                "user.type" to if (isDevUser) "internal" else "external",
+                installStatusPropertyName to "Active"
             )
         )
     }
@@ -637,6 +661,9 @@ class ActivityMonitor(project: Project) : Disposable {
     }
 
     fun registerUserAction(action: String) {
+
+        PersistenceService.getInstance().setLastUserActionTimestamp()
+
         capture(
             "user-action",
             mapOf("action" to action)