+++ model/src/main/kotlin/org/digma/intellij/plugin/model/rest/insights/EPNPlusSpansInsight.kt
@@ -33,6 +33,7 @@ data class EPNPlusSpansInsight
         "lastCommitId",
         "deactivatedCommitId",
         "reopenCount",
+        "ticketLink"
 )
 constructor(
         override val codeObjectId: String,
@@ -58,6 +59,7 @@ constructor(
         override val lastCommitId: String?,
         override val deactivatedCommitId: String?,
         override val reopenCount: Int,
+        override val ticketLink: String?,
 ) : EndpointInsight {
     override val type: InsightType = InsightType.EndpointSpaNPlusOne
 }