+++ model/src/main/kotlin/org/digma/intellij/plugin/model/rest/insights/SpanDurationsInsight.kt
@@ -29,6 +29,7 @@ data class SpanDurationsInsight
     "lastCommitId",
     "deactivatedCommitId",
     "reopenCount",
+    "ticketLink"
 )
 constructor(
     override val codeObjectId: String,
@@ -48,6 +49,7 @@ constructor(
     override val lastCommitId: String?,
     override val deactivatedCommitId: String?,
     override val reopenCount: Int,
+    override val ticketLink: String?,
 ) : SpanInsight {
 
     override val type: InsightType = InsightType.SpanDurations