+++ model/src/main/kotlin/org/digma/intellij/plugin/model/rest/insights/SpanSlowEndpointsInsight.kt
@@ -31,6 +31,7 @@ data class SpanSlowEndpointsInsight
         "lastCommitId",
         "deactivatedCommitId",
         "reopenCount",
+        "ticketLink"
 )
 constructor(
         override val codeObjectId: String,
@@ -54,6 +55,7 @@ constructor(
         override val lastCommitId: String?,
         override val deactivatedCommitId: String?,
         override val reopenCount: Int,
+        override val ticketLink: String?,
 ) : SpanInsight {
     override val type: InsightType = InsightType.SpanEndpointBottleneck
 }