+++ model/src/main/kotlin/org/digma/intellij/plugin/model/rest/insights/UnlinkTicketRequest.kt
@@ -1,0 +1,11 @@
+package org.digma.intellij.plugin.model.rest.insights
+
+import com.fasterxml.jackson.annotation.JsonCreator
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties
+import java.beans.ConstructorProperties
+
+data class UnlinkTicketRequest
+@JsonCreator(mode = JsonCreator.Mode.PROPERTIES)
+@JsonIgnoreProperties(ignoreUnknown = true)
+@ConstructorProperties("environment", "codeObjectId", "insightType")
+constructor(val environment: String, val codeObjectId: String, val insightType: String)