+++ src/main/java/org/digma/intellij/plugin/insights/InsightsMessageRouterHandler.java
@@ -20,6 +20,7 @@ import org.digma.intellij.plugin.analytics.AnalyticsService;
 import org.digma.intellij.plugin.analytics.AnalyticsServiceException;
 import org.digma.intellij.plugin.common.Backgroundable;
 import org.digma.intellij.plugin.common.EDT;
+import org.digma.intellij.plugin.dashboard.outgoing.BackendInfoMessage;
 import org.digma.intellij.plugin.document.CodeObjectsUtil;
 import org.digma.intellij.plugin.errorreporting.ErrorReporter;
 import org.digma.intellij.plugin.insights.model.outgoing.CommitInfo;
@@ -30,6 +31,7 @@ import org.digma.intellij.plugin.insights.model.outgoing.SetCodeLocationMessage;
 import org.digma.intellij.plugin.insights.model.outgoing.SetCommitInfoData;
 import org.digma.intellij.plugin.insights.model.outgoing.SetCommitInfoMessage;
 import org.digma.intellij.plugin.insights.model.outgoing.SetInsightsDataMessage;
+import org.digma.intellij.plugin.insights.model.outgoing.SetLinkUnlinkResponseMessage;
 import org.digma.intellij.plugin.insights.model.outgoing.SetSpanInsightData;
 import org.digma.intellij.plugin.insights.model.outgoing.SetSpanInsightMessage;
 import org.digma.intellij.plugin.insights.model.outgoing.Span;
@@ -37,6 +39,7 @@ import org.digma.intellij.plugin.jcef.common.JCefBrowserUtil;
 import org.digma.intellij.plugin.jcef.common.JCefMessagesUtils;
 import org.digma.intellij.plugin.log.Log;
 import org.digma.intellij.plugin.model.InsightType;
+import org.digma.intellij.plugin.model.rest.AboutResult;
 import org.digma.intellij.plugin.model.rest.insights.CodeObjectInsight;
 import org.digma.intellij.plugin.model.rest.insights.InsightsOfSingleSpanResponse;
 import org.digma.intellij.plugin.model.rest.jcef.common.SendTrackingEventRequest;
@@ -56,6 +59,7 @@ import org.jetbrains.annotations.NotNull;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.TreeMap;
 import java.util.stream.Collectors;
 
@@ -95,8 +99,7 @@ class InsightsMessageRouterHandler extends CefMessageRouterHandlerAdapter {
                 String action = jsonNode.get("action").asText();
                 switch (action) {
 
-                    case "INSIGHTS/INITIALIZE" -> {
-                    }
+                    case "INSIGHTS/INITIALIZE" -> onInitialize(browser);
 
                     case "INSIGHTS/GET_DATA" -> pushInsightsFromGetData();
 
@@ -125,6 +128,10 @@ class InsightsMessageRouterHandler extends CefMessageRouterHandlerAdapter {
                     case "INSIGHTS/ADD_ANNOTATION" -> addAnnotation(jsonNode);
 
                     case "INSIGHTS/MARK_INSIGHT_TYPES_AS_VIEWED" -> markInsightsViewed(jsonNode);
+
+                    case "INSIGHTS/LINK_TICKET" -> linkTicket(jsonNode);
+
+                    case "INSIGHTS/UNLINK_TICKET" -> unlinkTicket(jsonNode);
 
                     case JCefMessagesUtils.GLOBAL_OPEN_TROUBLESHOOTING_GUIDE ->
                             EDT.ensureEDT(() -> MainToolWindowCardsController.getInstance(project).showTroubleshooting());
@@ -201,6 +208,29 @@ class InsightsMessageRouterHandler extends CefMessageRouterHandlerAdapter {
         });
         Log.log(LOGGER::trace, project, "got insights types {}", insightTypeList);
         ActivityMonitor.getInstance(project).registerInsightsViewed(insightTypeList);
+    }
+
+    private void linkTicket(JsonNode jsonNode) throws JsonProcessingException, AnalyticsServiceException {
+        Log.log(LOGGER::trace, project, "got INSIGHTS/LINK_TICKET message");
+        var payload = objectMapper.readTree(jsonNode.get("payload").toString());
+        var codeObjectId = payload.get("codeObjectId").asText();
+        var insightType = payload.get("insightType").asText();
+        var ticketLink = payload.get("ticketLink").asText();
+        var linkTicketResponse = AnalyticsService.getInstance(project).linkTicket(codeObjectId, insightType, ticketLink);
+        var message = new SetLinkUnlinkResponseMessage("digma", "INSIGHTS/SET_TICKET_LINK", linkTicketResponse);
+        serializeAndExecuteWindowPostMessageJavaScript(this.jbCefBrowser.getCefBrowser(), message);
+        ActivityMonitor.getInstance(project).registerUserActionEvent("link ticket", Map.of("insight", insightType));
+    }
+
+    private void unlinkTicket(JsonNode jsonNode) throws JsonProcessingException, AnalyticsServiceException {
+        Log.log(LOGGER::trace, project, "got INSIGHTS/UNLINK_TICKET message");
+        var payload = objectMapper.readTree(jsonNode.get("payload").toString());
+        var codeObjectId = payload.get("codeObjectId").asText();
+        var insightType = payload.get("insightType").asText();
+        var unlinkTicketResponse = AnalyticsService.getInstance(project).unlinkTicket(codeObjectId, insightType);
+        var message = new SetLinkUnlinkResponseMessage("digma", "INSIGHTS/SET_TICKET_LINK", unlinkTicketResponse);
+        serializeAndExecuteWindowPostMessageJavaScript(this.jbCefBrowser.getCefBrowser(), message);
+        ActivityMonitor.getInstance(project).registerUserActionEvent("unlink ticket", Map.of("insight", insightType));
     }
 
     private CodeObjectInsight getInsightBySpanTemporary(String spanCodeObjectId, String insightType) throws AnalyticsServiceException {
@@ -391,6 +421,20 @@ class InsightsMessageRouterHandler extends CefMessageRouterHandlerAdapter {
     private void pushInsightsFromGetData() {
         Log.log(LOGGER::debug, project, "got INSIGHTS/GET_DATA message");
         InsightsService.getInstance(project).refreshInsights();
+    }
+
+    private void onInitialize(CefBrowser browser) {
+        try {
+            AboutResult about = AnalyticsService.getInstance(project).getAbout();
+            var message = new BackendInfoMessage(
+                    JCefMessagesUtils.REQUEST_MESSAGE_TYPE, JCefMessagesUtils.GLOBAL_SET_BACKEND_INFO,
+                    about);
+
+            Log.log(LOGGER::trace, project, "sending {} message",JCefMessagesUtils.GLOBAL_SET_BACKEND_INFO);
+            serializeAndExecuteWindowPostMessageJavaScript(browser, message);
+        } catch (AnalyticsServiceException e) {
+            Log.warnWithException(LOGGER, e, "error getting backend info");
+        }
     }