+++ src/main/kotlin/org/digma/intellij/plugin/ui/dashboard/DashboardButton.kt
@@ -5,6 +5,7 @@ import com.intellij.openapi.project.Project
 import com.intellij.ui.JBColor
 import com.intellij.util.ui.JBUI
 import org.digma.intellij.plugin.analytics.EnvironmentChanged
+import org.digma.intellij.plugin.common.adjustEnvironmentDisplayName
 import org.digma.intellij.plugin.dashboard.DashboardService
 import org.digma.intellij.plugin.errorreporting.ErrorReporter
 import org.digma.intellij.plugin.log.Log
@@ -33,7 +34,7 @@ class DashboardButton(val project: Project) : JButton() {
 
 
         //on fresh install there is no env yet
-        if (PersistenceService.getInstance().state.currentEnv == null) {
+        if (PersistenceService.getInstance().getCurrentEnv() == null) {
             isEnabled = false
             toolTipText = "No Environment Yet"
             project.messageBus.connect().subscribe(
@@ -41,12 +42,12 @@ class DashboardButton(val project: Project) : JButton() {
                 object : EnvironmentChanged {
 
                     override fun environmentChanged(newEnv: String?, refreshInsightsView: Boolean) {
-                        isEnabled = PersistenceService.getInstance().state.currentEnv != null
+                        isEnabled = PersistenceService.getInstance().getCurrentEnv() != null
                         toolTipText = if (isEnabled) "Open Dashboard" else "No Environment Yet"
                     }
 
                     override fun environmentsListChanged(newEnvironments: MutableList<String>?) {
-                        isEnabled = PersistenceService.getInstance().state.currentEnv != null
+                        isEnabled = PersistenceService.getInstance().getCurrentEnv() != null
                         toolTipText = if (isEnabled) "Open Dashboard" else "No Environment Yet"
                     }
                 })
@@ -62,7 +63,11 @@ class DashboardButton(val project: Project) : JButton() {
 
         try {
 
-            DashboardService.getInstance(project).openDashboard("Dashboard Panel - " + PersistenceService.getInstance().state.currentEnv);
+            val envName = PersistenceService.getInstance().getCurrentEnv()?.let {
+                adjustEnvironmentDisplayName(it)
+            } ?: "No Environment"
+
+            DashboardService.getInstance(project).openDashboard("Dashboard Panel - $envName");
 
         } catch (e: Exception) {
             Log.warnWithException(logger, project, e, "Error in doActionListener")